separating timeline component from traceDetails
